+++ lib/src/utils/extensions.dart
@@ -14,6 +14,7 @@ extension NumExtension on num {
 
 extension DoubleExtension on double {
   /// Returns the value linearly transformed from [valueMin, valueMax] to [0, 1].
+  /// If the value is not in the the given [valueMin, valueMax] it will be clamped to still fall in [0, 1].
   double normalize(num valueMin, num valueMax) => distribute(
         targetMin: 0,
         targetMax: 1,
@@ -22,6 +23,7 @@ extension DoubleExtension on double {
       );
 
   /// Returns the value linearly transformed from [valueMin, valueMax] to [targetMin, targetMax].
+  /// If the value is not in the the given [valueMin, valueMax] it will be clamped to still fall in [targetMin, targetMax].
   /// Source: https://stats.stackexchange.com/questions/178626/how-to-normalize-data-between-1-and-1
   double distribute({
     required num targetMin,
@@ -30,7 +32,9 @@ extension DoubleExtension on double {
     required num valueMax,
   }) {
     assert(valueMin != valueMax);
-    return (targetMax - targetMin) * ((this - valueMin) / (valueMax - valueMin)) + targetMin;
+    return ((targetMax - targetMin) * ((this - valueMin) / (valueMax - valueMin)) + targetMin)
+        .clamp(targetMin, targetMax)
+        .toDouble();
   }
 }