+++ lib/src/widgets/biomarker_card/biomarker_card_range_indicator.dart
@@ -51,9 +51,9 @@ class BiomarkerCardRangeIndicator extends StatelessWidget {
     if (biomarker.maxOptimalValue == null && biomarker.minOptimalValue == null) {
       text = '--';
     } else if (biomarker.maxOptimalValue == null) {
-      text = '> ${biomarker.minOptimalValue!.toStringPretty()}';
+      text = 'Over ${biomarker.minOptimalValue!.toStringPretty()}';
     } else if (biomarker.minOptimalValue == null) {
-      text = '< ${biomarker.maxOptimalValue!.toStringPretty()}';
+      text = 'Under ${biomarker.maxOptimalValue!.toStringPretty()}';
     } else {
       text = '${biomarker.minOptimalValue!.toStringPretty()} - ${biomarker.maxOptimalValue!.toStringPretty()}';
     }